refactor: adjust nested class (#50)
