+++ src/main/java/scanner/controller/GithubAppController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import scanner.response.CommonResponse;
+import scanner.dto.CommonResponse;
 import scanner.service.github.GithubAppService;
