--- src/main/java/scanner/dto/CheckListSimpleDto.java
@@ -1,46 +1,0 @@
-package scanner.dto;
-
-
-import lombok.*;
-import scanner.model.CustomRule;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CheckListSimpleDto {
-
-    private String ruleId;
-    private String ruleOnOff;
-
-    @Setter
-    private CheckListCustomDto custom;
-
-
-    public CheckListSimpleDto(final CustomRule rule){
-        this.ruleId = rule.getRuleId();
-        this.ruleOnOff = rule.getRuleOnOff();
-        this.custom = new CheckListCustomDto(rule.getCustomDetail());
-    }
-
-    public static CheckListSimpleDto toDto(CustomRule rule) {
-        return CheckListSimpleDto.builder()
-                                .ruleId(rule.getRuleId())
-                                .ruleOnOff(rule.getRuleOnOff())
-                                .custom(new CheckListCustomDto(rule.getCustomDetail()))
-                                .build();
-    }
-
-    public static CustomRule toEntity(final CheckListSimpleDto dto){
-        if(dto.getCustom() == null)
-            return CustomRule.builder()
-                            .ruleOnOff(dto.getRuleOnOff())
-                            .customDetail(null)
-                            .build();
-
-        return CustomRule.builder()
-                        .ruleOnOff(dto.getRuleOnOff())
-                        .customDetail(dto.getCustom().getCustomDetail())
-                        .build();
-    }
-}
