+++ src/main/java/scanner/dto/ExceptionDto.java
@@ -3,7 +3,7 @@ package scanner.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import scanner.response.enums.ResponseCode;
+import scanner.dto.enums.ResponseCode;
 
 @AllArgsConstructor
 @Getter
