--- src/main/java/scanner/dto/ParseResultDto.java
@@ -1,13 +1,0 @@
-package scanner.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ParseResultDto {
-    Object result;
-}
