--- src/main/java/scanner/dto/SolutionDto.java
@@ -1,13 +1,0 @@
-package scanner.dto;
-
-
-import lombok.*;
-
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class SolutionDto {
-    private String sol;
-    private String code;
-}
