+++ src/main/java/scanner/dto/checklist/CheckListCustomDto.java
@@ -1,4 +1,4 @@
-package scanner.dto;
+package scanner.dto.checklist;
 
 import lombok.*;
