+++ src/main/java/scanner/dto/enums/ResponseCode.java
@@ -1,4 +1,4 @@
-package scanner.response.enums;
+package scanner.dto.enums;
 
 import org.springframework.http.HttpStatus;
