+++ src/main/java/scanner/dto/report/FailedComplianceDto.java
@@ -1,4 +1,4 @@
-package scanner.dto.history.report;
+package scanner.dto.report;
 
 import java.util.LinkedHashMap;
 import java.util.List;
