+++ src/main/java/scanner/dto/report/FailedDto.java
@@ -1,4 +1,4 @@
-package scanner.dto.history.report;
+package scanner.dto.report;
 
 import java.util.ArrayList;
 import java.util.List;
