+++ src/main/java/scanner/dto/report/ScanSummaryDto.java
@@ -1,4 +1,4 @@
-package scanner.dto.history.report;
+package scanner.dto.report;
 
 import java.time.format.DateTimeFormatter;
 import java.util.List;
