+++ src/main/java/scanner/exception/ApiException.java
@@ -3,7 +3,7 @@ package scanner.exception;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import scanner.response.enums.ResponseCode;
+import scanner.dto.enums.ResponseCode;
 
 @AllArgsConstructor
 @RequiredArgsConstructor
