+++ src/main/java/scanner/exception/handler/CommonExceptionHandler.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import scanner.exception.ApiException;
-import scanner.response.CommonResponse;
+import scanner.dto.CommonResponse;
 
 @RestControllerAdvice
 @Slf4j
