+++ src/main/java/scanner/security/service/CustomUserDetailService.java
@@ -10,7 +10,7 @@ import scanner.dto.user.UserRetrieveDto;
 import scanner.exception.ApiException;
 import scanner.model.User;
 import scanner.repository.UserRepository;
-import scanner.response.enums.ResponseCode;
+import scanner.dto.enums.ResponseCode;
 
 import javax.persistence.*;
 import javax.transaction.Transactional;
